new gradual file
