+++ liquid-fixpoint.cabal
@@ -104,6 +104,7 @@ Library
                    Language.Fixpoint.Solver.Eliminate,
                    Language.Fixpoint.Solver.Sanitize,
                    Language.Fixpoint.Utils.Statistics,
+                   Language.Fixpoint.Solver.GradualSolve,
                    Language.Fixpoint.Solver.Solve
 
   Build-Depends: base >= 4.8 && < 5
