+++ src/Language/Fixpoint/Solver/Monad.hs
@@ -174,8 +174,6 @@ filterRequired = error "TBD:filterRequired"
 > unsat (b2 bR)
 -}
 
-
-
 --------------------------------------------------------------------------------
 -- | `filterValid p [(x1, q1),...,(xn, qn)]` returns the list `[ xi | p => qi]`
 --------------------------------------------------------------------------------
